backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "coverage>=7.13.5",
    "fastapi[standard]>=0.135.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "uuid6>=2025.0.1",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I"]
